@@ -2,9 +2,9 @@
 {
 	public class AppSettings
 	{
-		public string InstallersFolder { get; set; }
-		public string RepoFolder { get; set; }
-		public string[] LocalesToDownload { get; set; }
+		public string InstallersFolder { get; set; } = "";
+		public string RepoFolder { get; set; } = "";
+		public string[] LocalesToDownload { get; set; } = [];
 
-	}
+    }
 }
